+++ app/components/airdrops/cards/AirdropInfo.tsx
@@ -1,13 +1,10 @@
 import { Box, Text } from "@chakra-ui/react";
 import { AirdropInfoProps } from "@/app/interfaces/dashboard";
 import { useTranslation } from "react-i18next";
-import {
-  formatTemplateType
-} from "@/app/lib/airdrop/airdropUtils";
+import { formatTemplateType } from "@/app/lib/airdrop/airdropUtils";
 
 export default function AirdropInfo({ airdropTitle, templateType }: AirdropInfoProps) {
   const { t } = useTranslation();
-  console.log(t(formatTemplateType(templateType)))
   return (
     <>
       <Text fontSize={{ base: "25px", sm: "30px" }} fontWeight={{ base: "600", sm: "400" }}>
